reinforcement-learning/deep-rts/requirements.txt
numpy
flask
flask-socketio
socketIO-client
peewee
DeepRTS
eventlet
tqdm